+++ ex03/src/Fixed.cpp
@@ -6,7 +6,7 @@
 /*   By: hfilipe- <hfilipe-@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/16 11:39:26 by hfilipe-          #+#    #+#             */
-/*   Updated: 2025/05/19 14:32:40 by hfilipe-         ###   ########.fr       */
+/*   Updated: 2025/05/20 19:26:39 by hfilipe-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ Fixed::Fixed() : fixedNb(0) {
 
 Fixed::Fixed(const int value) : fixedNb(value << bits) {}
 
-Fixed::Fixed(const float value) : fixedNb(static_cast<int>(round(value * (1 << bits)))) {}
+Fixed::Fixed(const float value) : fixedNb(static_cast<int>(roundf(value * (1 << bits)))) {}
 
 Fixed::Fixed(const Fixed &other) : fixedNb(other.fixedNb){
 	std::cout << "Copy constructor called" << std::endl;
